[V3.0] See desrciption for changelog (it's a big one)
[V3.0]

- Introduced delete command messages. Can be toggled on and off by administrators using the "deletecommandmessages" (alias "dcm") command

- Introduced toggling modlogs on and off. Can be toggled with the "modlogs" (alias "togglemod") command. Defaults to on

- Introduce the "hasSentModLogMessage" parameters which ensures the notifier concering modlogs is only sent 1x per server

- Introduce configurable modlogs channel with the "setmodlogs" (alias "setmod") command. Requires modlogs to be on for it to fully work

- Introduce configurable memberlogs channel with the "setmemberlogs" (alias "setmember") command. Requires memberlogs to be on for it to fully work

- Various commands have been rewritten to be asyncronous using async/await functions: strawpoll, weather, oxr, all nsfw commands, dictionary define, anime, google, google images and steam.

- google command has been fixed, previously knowledge graph results would always error and be completely unused.

- various response and error messages have been changed.

- lots of performance tweaks such as not defining extra constants, defining methods within classes instead of outside, using argument parsing instead of functions and so on.

- games command has been completely rewritten from scratch. Instead of using mobygames it now uses IGDB (IndieGames DB) which despite the name actually has info on pretty much all games and has a great API to boot.

- Removal of the superagent and request packages as I now use snekfetch everywhere

- Adding of the igdb-api-node package for the IGDB command

+++ src/commands/misc/say.js
@@ -62,11 +62,13 @@ module.exports = class sayCommand extends commando.Command {
 		});
 	}
 
-	run (msg, args) {
-		if (msg.deletable) {
+	deleteCommandMessages (msg) {
+		if (msg.deletable && this.client.provider.get(msg.guild, 'deletecommandmessages', false)) {
 			msg.delete();
 		}
+	}
 
+	run (msg, args) {
 		let storageLooped = false;
 
 		for (const stored in storage.lastMessage) {
@@ -80,7 +82,8 @@ module.exports = class sayCommand extends commando.Command {
 		if (!storageLooped) {
 			storage.lastMessage.push([msg.guild.id, {'message': msg}]);
 		}
-	
+		this.deleteCommandMessages(msg);
+
 		return msg.say(args.txt);
 	}
 };
